backend/requirements.txt
# FastAPI and server
fastapi==0.110.0
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9

# Redis
redis==5.0.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# AI
google-genai==0.3.0

# File handling
python-magic==0.4.27
PyPDF2==3.0.1
python-docx==1.1.0
Pillow==10.2.0

# OCR (optional)
pytesseract==0.3.10

# Utilities
pydantic==2.6.1
pydantic-settings==2.1.0
python-dotenv==1.0.1
httpx==0.26.0
aiofiles==23.2.1

# Development
pytest==8.0.0
pytest-asyncio==0.23.4
